+++ qubes_menu/appmenu.py
@@ -15,7 +15,6 @@ import logging
 import qubesadmin
 import qubesadmin.events
 
-from .settings_page import SettingsPage
 from .application_page import AppPage
 from .desktop_file_manager import DesktopFileManager
 from .favorites_page import FavoritesPage
@@ -59,18 +58,13 @@ class AppMenu(Gtk.Application):
         self.main_window: Optional[Gtk.Window] = None
         self.main_notebook: Optional[Gtk.Notebook] = None
 
-        self.fav_app_list: Optional[Gtk.ListBox] = None
-        self.sys_tools_list: Optional[Gtk.ListBox] = None
-
         self.desktop_file_manager: Optional[DesktopFileManager] = None
         self.vm_manager: Optional[VMManager] = None
         self.app_page: Optional[AppPage] = None
 
         self.favorites_page: Optional[FavoritesPage] = None
-        self.settings_page: Optional[SettingsPage] = None
 
         self.power_button: Optional[Gtk.Button] = None
-        self.tasks = []
 
     def _add_cli_options(self):
         self.add_main_option(
@@ -165,6 +159,78 @@ class AppMenu(Gtk.Application):
                 else:
                     self.main_window.present()
 
+
+    def initialize_state(self):
+        """
+        Initial state, that is - menu is open at the 0th page and pages
+        will initialize their state if needed. Separate function because
+        some things (like widget size adjustments) must be called after
+        widgets are realized and not on init.
+        """
+        if self.main_notebook:
+            self.main_notebook.set_current_page(self.initial_page)
+        if self.app_page:
+            self.app_page.initialize_state(None)
+
+    def perform_setup(self):
+        """
+        The function that performs actual widget realization and setup. Should
+        be only called once, in the main instance of this application.
+        """
+        screen = Gdk.Screen.get_default()
+
+        provider = Gtk.CssProvider()
+        provider.load_from_path(
+            pkg_resources.resource_filename(__name__, 'qubes-menu-dark.css')
+        )
+
+        Gtk.StyleContext.add_provider_for_screen(
+            screen, provider, Gtk.STYLE_PROVIDER_PRIORITY_APPLICATION
+        )
+
+        self.builder = Gtk.Builder()
+
+        self.builder.add_from_file(
+            pkg_resources.resource_filename(__name__, 'qubes-menu.glade')
+        )
+
+        self.main_window = self.builder.get_object('main_window')
+        self.main_notebook = self.builder.get_object('main_notebook')
+        self.power_button = self.builder.get_object('power_button')
+        
+        self.desktop_file_manager = DesktopFileManager(self.qapp)
+        
+        self.vm_manager = VMManager(self.qapp, self.dispatcher)
+
+        self.app_page = AppPage(
+            self.vm_manager, self.builder, self.desktop_file_manager, self.dispatcher
+        )
+        
+        self.favorites_page = FavoritesPage(
+            self.qapp, self.builder, self.desktop_file_manager, self.dispatcher, self.vm_manager
+        )
+
+        self.add_window(self.main_window)
+        
+
+        self.main_window.set_events(Gdk.EventMask.FOCUS_CHANGE_MASK)
+        self.main_window.connect('focus-out-event', self._focus_out)
+        self.main_window.connect('key_press_event', self._key_press)
+        
+        self.main_notebook.connect('switch-page', self._handle_page_switch)
+        
+        self.power_button.connect('clicked', self._do_power_button)
+        
+        self.connect('shutdown', self.do_shutdown)
+
+    def _handle_page_switch(self, _widget, _page, page_num):
+        """
+        On page switch some things need to happen, mostly cleaning any old
+        selections/menu options highlighted.
+        """
+        if page_num == 0 and self.app_page:
+            self.app_page.initialize_state()
+
     def hide_menu(self):
         """
         Unless CLI options specified differently, the menu will try to hide
@@ -187,69 +253,7 @@ class AppMenu(Gtk.Application):
         """
         if SelfAwareMenu.OPEN_MENUS <= 0:
             self.hide_menu()
-
-    def initialize_state(self):
-        """
-        Initial state, that is - menu is open at the 0th page and pages
-        will initialize their state if needed. Separate function because
-        some things (like widget size adjustments) must be called after
-        widgets are realized and not on init.
-        """
-        if self.main_notebook:
-            self.main_notebook.set_current_page(self.initial_page)
-        if self.app_page:
-            self.app_page.initialize_state(None)
-
-    def perform_setup(self):
-        """
-        The function that performs actual widget realization and setup. Should
-        be only called once, in the main instance of this application.
-        """
-        screen = Gdk.Screen.get_default()
-        provider = Gtk.CssProvider()
-        provider.load_from_path(pkg_resources.resource_filename(
-            __name__, 'qubes-menu-dark.css'))
-        Gtk.StyleContext.add_provider_for_screen(
-            screen, provider, Gtk.STYLE_PROVIDER_PRIORITY_APPLICATION)
-        self.builder = Gtk.Builder()
-
-        self.fav_app_list = self.builder.get_object('fav_app_list')
-        self.sys_tools_list = self.builder.get_object('sys_tools_list')
-        self.builder.add_from_file(pkg_resources.resource_filename(
-            __name__, 'qubes-menu.glade'))
-        self.main_window = self.builder.get_object('main_window')
-        self.main_notebook = self.builder.get_object('main_notebook')
-
-        self.main_window.set_events(Gdk.EventMask.FOCUS_CHANGE_MASK)
-        self.main_window.connect('focus-out-event', self._focus_out)
-        self.main_window.connect('key_press_event', self._key_press)
-        self.add_window(self.main_window)
-        self.desktop_file_manager = DesktopFileManager(self.qapp)
-        self.vm_manager = VMManager(self.qapp, self.dispatcher)
-
-        self.app_page = AppPage(self.vm_manager, self.builder,
-                                self.desktop_file_manager, self.dispatcher)
-        self.favorites_page = FavoritesPage(self.qapp, self.builder,
-                                            self.desktop_file_manager,
-                                            self.dispatcher, self.vm_manager)
-        self.power_button = self.builder.get_object('power_button')
-        self.power_button.connect('clicked', self._do_power_button)
-        self.main_notebook.connect('switch-page', self._handle_page_switch)
-        self.connect('shutdown', self.do_shutdown)
-
-        self.tasks = [
-            asyncio.ensure_future(self.dispatcher.listen_for_events())]
-
-    def _handle_page_switch(self, _widget, _page, page_num):
-        """
-        On page switch some things need to happen, mostly cleaning any old
-        selections/menu options highlighted.
-        """
-        if page_num == 0 and self.app_page:
-            self.app_page.initialize_state()
-        elif page_num == 2 and self.settings_page:
-            self.settings_page.initialize_state()
-
+            
     def exit_app(self):
         """
         Exit. Used by restart only at the moment, as the menu is designed to
